fixed music support for linux
Earlier, due to lack of support of music for linux, the application used to crash. Now, the user can modify the 'os_supports_music' configuration variable on top of the Tetris.py file.

If os_supports_music is disabled, even in a supported OS like windows, the music is not loaded.